chore: move the escape sequence functions
