+++ src/displays/app_description.c
@@ -1,5 +1,4 @@
 #include "../utils/utils.h"
-#include "escape_sequence.h"
 
 void app_description(void) {
   add_new_tab(1);
