--- src/displays/escape_sequence.h
@@ -1,8 +1,0 @@
-#ifndef ESCAPE_SEQUENCE_H
-#define ESCAPE_SEQUENCE_H
-#include "escape_sequence.c"
-
-extern void add_new_line(int num_lines);
-extern void add_new_tab(int num_tabs);
-
-#endif
