+++ src/scales/celcius_description.c
@@ -1,5 +1,4 @@
 #include "../utils/utils.h"
-#include "../displays/escape_sequence.h"
 
 void celsius_description(void) {
   add_new_tab(1);
