+++ src/utils/add_new_line.c
@@ -5,9 +5,3 @@ void add_new_line(int num_lines) {
     printf("\n");
   }
 }
-
-void add_new_tab(int num_tabs) {
-  for (int i = 0; i < num_tabs; i++) {
-    printf("\t");
-  }
-}
