+++ src/utils/add_new_tab.c
@@ -1,0 +1,7 @@
+#include <stdio.h>
+
+void add_new_tab(int num_tabs) {
+  for (int i = 0; i < num_tabs; i++) {
+    printf("\t");
+  }
+}
